Implement build --ignore-failures:
Because we try to maintain a transaction for all the packages we are
asked to build, if one of them fails, it causes a failure to all.

If you're building a lot of packages and you know some of the will fail,
you can now use the "--ignore-failures" flag, which will simply continue
with the next build when it fails.

However, this does not continue with dependencies if it fails.