Revamp command line interface for "manage" command:
I've hated this for a while. It grew in complexity over time and we
never made sure it stays consistent. I think this is a step in the right
direction.

Previously,

* "add", "remove", and "show" were for the package, implicitly.
* Some commands had "list", others add "show".
* Some commands had underscores ("remove_parcel").
* There were sub-subcommands ("pakket manage list parcels").
* Dependency commands were *really* hard to fathom and remember.

Now,

* All commands are either prefixed with "add", "remove", or "list".
* All commands have two parts: verb ("add", "remove", "list") and
  object ("package", "parcel", "deps").
* Dependencies now declare the dependency name using "--on" and the
  phase using "--phase".

That's it.

This is a bit tricky because it includes a layer that translates the
previous interface to the new one so it should work too. No tests, so
good luck to us all. :)