Fix installer --ignore-failures:
There is a problem with --ignore-failures where it will reuse the same
Pakket::Installer instance, which already includes a work directory
which will be removed at the end of an installation.

So:

* Create work dir
* Install single package
* Remove work dir
* Install single package

Thus, we have to creat the Installer instance every time. Hence this
commit.

It also includes a small compilation problem where a variable was not
imported so it failed. Sorry.

There is a better way to fix the installer being able to break on
packages which I will try to do in the future.